API: Move serviceinfo to core
Best to expose this endpoint to all APIs as its an information endpoint.
